Merge branch Rafael

+++ src/main/java/web/onficina/model/AvaliacaoReportDTO.java
@@ -2,7 +2,7 @@ package web.onficina.model;
 
 public class AvaliacaoReportDTO {
 
-    private Integer nota;
+    private Double nota;
     private String comentario;
     private String nomeCliente; 
 
@@ -13,7 +13,7 @@ public class AvaliacaoReportDTO {
     }
 
     // Getters
-    public Integer getNota() {
+    public Double getNota() {
         return nota;
     }
